requirements.txt
django
djangorestframework
drf-writable-nested
djangorestframework-simplejwt
algoliasearch-django>=3.0,<4.0